requirements.txt
torch
torchvision
scikit-learn
seaborn
numpy
scikit-image
matplotlib
tqdm
pillow
